+++ python/IECoreArnoldTest/ParameterAlgoTest.py
@@ -187,3 +187,19 @@ class ParameterAlgoTest( unittest.TestCase ) :
 			self.assertEqual( len( mh.messages ), 2 )
 			self.assertEqual( mh.messages[0].message, 'Int64Data value 2147483648 is out of range for parameter "customInt64OutOfRange"' )
 			self.assertEqual( mh.messages[1].message, 'UInt64Data value 4294967296 is out of range for parameter "customUInt64OutOfRange"' )
+
+	def testV3dVectorData( self ) :
+
+		with IECoreArnold.UniverseBlock( writable = True ) as universe :
+
+			n = arnold.AiNode( universe, "ginstance" )
+			IECoreArnold.ParameterAlgo.setParameter( n, "test", IECore.V3dVectorData( [ imath.V3d( i, i + 1, i + 2 ) for i in range( 0, 10 ) ] ) )
+
+			a = arnold.AiNodeGetArray( n, "test" )
+			self.assertEqual( arnold.AiArrayGetType( a ), arnold.AI_TYPE_VECTOR )
+			self.assertEqual( arnold.AiArrayGetNumElements( a ), 10 )
+			for i in range( 0, 10 ) :
+				self.assertEqual(
+					arnold.AiArrayGetVec( a, i ),
+					arnold.AtVector( i, i + 1, i + 2 )
+				)