+++ Changes.md
@@ -29,6 +29,7 @@ Improvements
 
 - SetFilter, Scene Editors : Set expressions containing operators are now editable via drag and drop of set names.
 - NodeEditor : Scene locations can now be dropped to pin the editor to their source node. Locations may be dragged from the Viewer, AttributeEditor, LightEditor, HierarchyView and LightLinkingEditor.
+- Arnold : Added support for V3dVectorData primitive variables. These are converted to float vectors, since Arnold does not support doubles.
 
 Fixes
 -----
@@ -36,6 +37,11 @@ Fixes
 - LevelSetOffset : Fixed crash when attempting to offset grid types other than `FloatGrid`.
 - AttributeTweaks, CustomAttributes, OptionTweaks, CustomOptions, OptionQuery : Fixed contexts used by "From Scene", "From Selected" and "From Affected" menu items. This fixes errors caused by missing context variables (such as script variables).
 - GraphEditor : Stopped drag & dropped scene locations from navigating to private internal nodes.
+
+API
+---
+
+- ImageAlgo : Deprecated `image()` method. This will be removed in Gaffer 1.8.
 
 Documentation
 -------------